requirements.txt
mediapipe
opencv-python
pydantic
pyserial
numpy
argparse
